reverse a string

+++ 0x05-pointers_arrays_strings/5-rev_string.c
@@ -10,11 +10,11 @@ void rev_string(char *s)
 	int i;
 	char rev[strlen(s)];
 
-	for (i = 0 ; i < strlen(s) ; i++)
+	for (i = 0 ; i < (int) strlen(s) ; i++)
 	{
 		rev[strlen(s) - 1 - i] = s[i];
 	}
-	for (i = 0 ; i < strlen(s) ; i++)
+	for (i = 0 ; i < (int) strlen(s) ; i++)
 	{
 		s[i] = rev[i];
 	}
